+++ kivy_garden/frostedglass/__init__.py
@@ -13,262 +13,661 @@ __all__ = ('FrostedGlass', )
 
 from ._version import __version__
 
-from kivy.clock import Clock
-from kivy.metrics import dp
-from kivy.uix.floatlayout import FloatLayout
-from kivy.properties import (ListProperty, ObjectProperty, ColorProperty,
-                             NumericProperty)
-from kivy.graphics import (RenderContext, Fbo, Color, RoundedRectangle, Scale,
-                           Translate, PushMatrix, PopMatrix, ClearColor,
-                           ClearBuffers, BindTexture)
+from functools import partial
+from time import perf_counter as now
 
-shader_effect = '''
+from kivy.clock import Clock
+from kivy.core.window import Window
+from kivy.metrics import dp
+from kivy.graphics import (
+    BindTexture,
+    Color,
+    ClearColor,
+    ClearBuffers,
+    Fbo,
+    Rectangle,
+    RenderContext,
+    RoundedRectangle,
+    Scale,
+    SmoothLine,
+    Translate
+)
+from kivy.properties import (
+    ColorProperty,
+    ListProperty,
+    NumericProperty,
+    ObjectProperty,
+    StringProperty
+)
+from kivy.uix.floatlayout import FloatLayout
+
+# Used for type checking
+from kivy.uix.scrollview import ScrollView
+from kivy.uix.screenmanager import Screen
+from kivy.uix.image import Image
+
+
+MEAN_RES = (Window.width + Window.height)/2
+
+
+vertex_shader = ("""
+#ifdef GL_ES
+    precision highp float;
+#endif
+
+/* Outputs to the fragment shader */
+varying vec4 frag_color;
+varying vec2 tex_coord0;
+
+/* vertex attributes */
+attribute vec2     vPosition;
+attribute vec2     vTexCoords0;
+
+/* uniform variables */
+uniform mat4       modelview_mat;
+uniform mat4       projection_mat;
+uniform vec4       color;
+
+
+void main (void) {
+  frag_color = color;
+  tex_coord0 = vTexCoords0;
+  gl_Position = projection_mat * modelview_mat * vec4(vPosition.xy, 0.0, 1.0);
+}
+""")
+
+
+vertical_blur_shader = """
+#ifdef GL_ES
+precision lowp float;
+#endif
+
+/* Outputs from the vertex shader */
+varying vec4 frag_color;
+varying vec2 tex_coord0;
+
+/* uniform texture samplers */
+uniform sampler2D texture0;
+
+
+vec4 effect(vec4 color, sampler2D texture, vec2 tex_coords, vec2 coords)
+{{
+  float mean_res = float({});
+  float dt = (({} / 2.0) * 1.0 / mean_res);
+  vec4 sum = vec4(0.0);
+
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y - 3.0*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y - 2.5*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y - 2.0*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y - 1.5*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y - 1.0*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y - 0.5*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y + 0.5*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y + 1.0*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y + 1.5*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y + 2.0*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y + 2.5*dt)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y + 3.0*dt)) * 0.077;
+
+  return frag_color * vec4(sum.rgba);
+}}
+"""
+
+
+horizontal_blur_shader = """
+#ifdef GL_ES
+precision lowp float;
+#endif
+
+/* Outputs from the vertex shader */
+varying vec4 frag_color;
+varying vec2 tex_coord0;
+
+/* uniform texture samplers */
+uniform sampler2D texture0;
+
+
+vec4 effect(vec4 color, sampler2D texture, vec2 tex_coords, vec2 coords)
+{{
+  float mean_res = float({});
+  float dt = ({} / 2.0) * 1.0 / mean_res;
+  vec4 sum = vec4(0.0);
+
+
+  sum += texture2D(texture, vec2(tex_coords.x - 3.0*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x - 2.5*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x - 2.0*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x - 1.5*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x - 1.0*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x - 0.5*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x + 0.5*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x + 1.0*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x + 1.5*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x + 2.0*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x + 2.5*dt, tex_coords.y)) * 0.077;
+  sum += texture2D(texture, vec2(tex_coords.x + 3.0*dt, tex_coords.y)) * 0.077;
+
+  return  frag_color * vec4(sum.rgba);
+}}
+"""
+
+
+shader_footer_effect = """
+void main (void){
+  vec4 normal_color = frag_color * texture2D(texture0, tex_coord0);
+  vec4 effect_color = effect(
+      normal_color, texture0, tex_coord0, gl_FragCoord.xy
+  );
+  gl_FragColor = effect_color;
+}
+"""
+
+
+noise_shader = """
+#ifdef GL_ES
+precision highp float;
+#endif
+
+/* Outputs from the vertex shader */
+varying vec4 frag_color;
+varying vec2 tex_coord0;
+
+/* uniform texture samplers */
+uniform sampler2D texture0;
+
+
+vec4 effect(vec4 color, sampler2D texture, vec2 tex_coords, vec2 coords)
+{{
+    vec4 sum = vec4(0.0);
+
+    float a = 12.9898;
+    float b = 78.233;
+    float c = 43758.5453;
+    float dt2 = dot(tex_coords.xy, vec2(a, b));
+    float sn= mod(dt2, 3.14);
+    vec4 noise = vec4(fract(sin(sn) * c));
+
+    vec4 bg = texture2D(texture0, tex_coord0);
+    sum = mix(bg, vec4(noise.xyz, 1.0), 1.0);
+
+    return vec4(sum.rgba);
+}}
+"""
+
+
+final_shader_effect = """
+#ifdef GL_ES
+    precision highp float;
+#endif
+
 $HEADER$
 
-
+uniform int add_child_texture;
 uniform float opacity;
-uniform float blur_size;
 uniform float luminosity;
-uniform float noiseOpacity;
+uniform float saturation;
+uniform float noise_opacity;
 uniform vec2 position;
 uniform vec2 resolution;
 uniform vec4 color_overlay;
-uniform sampler2D bg;
+uniform sampler2D texture1;
+uniform sampler2D texture2;
 
+
+float noise(vec2 co)
+{
+    float a = 12.9898;
+    float b = 78.233;
+    float c = 43758.5453;
+    float dt= dot(co.xy, vec2(a, b));
+    float sn= mod(dt, 3.14);
+    return fract(sin(sn) * c);
+}
 
 void main(void)
 {
     vec2 pos = (gl_FragCoord.xy - position.xy) / resolution.xy;
-
     float noise = 0.0;
-    vec2 blurRadius = blur_size/resolution.xy;
-    vec3 noiseTexture = vec3(0.0);
     vec4 textureResult = vec4(0.0);
-    vec4 textureSum = texture2D(bg, pos);
+    vec4 effect_texture = texture2D(texture1, pos);
+    vec4 noise_texture = texture2D(texture2, tex_coord0);
 
-    textureSum += texture2D(bg, (pos + vec2(cos(0.00),sin(0.00)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.00),sin(0.00)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.00),sin(0.00)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.00),sin(0.00)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.63),sin(0.63)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.63),sin(0.63)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.63),sin(0.63)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(0.63),sin(0.63)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.26),sin(1.26)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.26),sin(1.26)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.26),sin(1.26)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.26),sin(1.26)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.88),sin(1.88)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.88),sin(1.88)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.88),sin(1.88)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(1.88),sin(1.88)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(2.51),sin(2.51)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(2.51),sin(2.51)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(2.51),sin(2.51)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(2.51),sin(2.51)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.14),sin(3.14)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.14),sin(3.14)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.14),sin(3.14)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.14),sin(3.14)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.77),sin(3.77)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.77),sin(3.77)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.77),sin(3.77)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(3.77),sin(3.77)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(4.40),sin(4.40)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(4.40),sin(4.40)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(4.40),sin(4.40)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(4.40),sin(4.40)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.03),sin(5.03)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.03),sin(5.03)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.03),sin(5.03)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.03),sin(5.03)) * blurRadius
-                            * 1.0));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.65),sin(5.65)) * blurRadius
-                            * 0.25));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.65),sin(5.65)) * blurRadius
-                            * 0.5));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.65),sin(5.65)) * blurRadius
-                            * 0.75));
-    textureSum += texture2D(bg, (pos + vec2(cos(5.65),sin(5.65)) * blurRadius
-                            * 1.0));
+    const vec3 W = vec3(0.2125, 0.7154, 0.0721);
+    vec3 intensity = vec3(dot(effect_texture.rgb, W));
+    effect_texture = vec4(mix(intensity, effect_texture.rgb, saturation), 1.0);
+    effect_texture *= luminosity;
 
-    textureSum /= 4.0 * 10.0 - 15.0;
-    textureSum *= luminosity;
+    effect_texture = mix(
+        effect_texture.rgba,
+        vec4(color_overlay.rgb, 1.0),
+        min(1.0, color_overlay.a)
+    );
+    effect_texture = mix(
+        effect_texture.rgba,
+        vec4(noise_texture.rgb, 1.0),
+        min(1.0, noise_opacity)
+    );
 
-    noise = fract(sin(dot((vec2(sin(pos) / 1.0)).xy, vec2(12.9898, 78.233)))
-                  * 43758.5453);
-    noiseTexture = vec3(noise);
-
-    textureSum = mix(textureSum.rgba, vec4(color_overlay.rgb, 1.0),
-                     color_overlay.a);
-    textureSum = mix(textureSum.rgba, vec4(noiseTexture.rgb, 1.0),
-                     noiseOpacity);
-
-    gl_FragColor = vec4(textureSum.rgb, opacity);
-
+    gl_FragColor = vec4(effect_texture.rgb, opacity);
 }
-'''
+"""
+
+
+class VerticalBlur(FloatLayout):
+
+    glsl = StringProperty('')
+    fbo = ObjectProperty(None, allownone=True)
+    blur_size = NumericProperty(0.0)
+
+    def __init__(self, *args, **kwargs):
+        super(VerticalBlur, self).__init__(*args, **kwargs)
+        self.bind(
+            fbo=self.set_fbo_shader,
+            blur_size=self.update_glsl,
+        )
+        self.glsl = horizontal_blur_shader.format(
+            MEAN_RES, float(self.blur_size)
+        )
+
+        with self.canvas:
+            self.rect = Rectangle()
+
+    def set_fbo_shader(self, *args):
+        if self.fbo is None:
+            return
+        self.fbo.shader.fs = self.glsl + shader_footer_effect
+
+    def update_glsl(self, *args):
+        self.glsl = horizontal_blur_shader.format(
+            MEAN_RES, float(self.blur_size)
+        )
+
+
+class HorizontalBlur(FloatLayout):
+
+    glsl = StringProperty('')
+    fbo = ObjectProperty(None, allownone=True)
+    blur_size = NumericProperty(0.0)
+
+    def __init__(self, *args, **kwargs):
+        super(HorizontalBlur, self).__init__(*args, **kwargs)
+        self.bind(
+            fbo=self.set_fbo_shader,
+            blur_size=self.update_glsl,
+        )
+        self.glsl = vertical_blur_shader.format(
+            MEAN_RES, float(self.blur_size)
+        )
+
+    def set_fbo_shader(self, *args):
+        if self.fbo is None:
+            return
+        self.fbo.shader.fs = self.glsl + shader_footer_effect
+
+    def update_glsl(self, *args):
+        self.glsl = vertical_blur_shader.format(
+            MEAN_RES, float(self.blur_size)
+        )
+
+
+class Noise(FloatLayout):
+
+    glsl = StringProperty('')
+    fbo = ObjectProperty(None, allownone=True)
+
+    def __init__(self, *args, **kwargs):
+        super(Noise, self).__init__(*args, **kwargs)
+        self.bind(fbo=self.set_fbo_shader)
+        with self.canvas:
+            self.rect = Rectangle()
+        self.glsl = noise_shader
+
+    def set_fbo_shader(self, *args):
+        if self.fbo is None:
+            return
+        self.fbo.shader.fs = self.glsl + shader_footer_effect
 
 
 class FrostedGlass(FloatLayout):
 
     background = ObjectProperty(None)
-    '''Target widget/layout that will be used as a background to FrostedGlass.
+    """Target widget/layout that will be used as a background to FrostedGlass.
     The recomended way to pass the widget is through the id.
 
     :attr:`background` is a :class:`~kivy.properties.ObjectProperty` and
-    defaults to None.'''
+    defaults to None."""
 
-    blur_size = NumericProperty(60)
-    '''Size of the gaussian blur aplied to the background.
+    blur_size = NumericProperty(25)
+    """Size of the gaussian blur aplied to the background.
+
+    Note: Do not pass relative values such as dp or sp. FrostedGlass already
+    manages this automatically, according to the device's screen density.
 
     :attr:`blur_size` is a :class:`~kivy.properties.NumericProperty` and
-    defaults to 60.'''
+    defaults to 25."""
 
     noise_opacity = NumericProperty(0.1)
-    '''Opacity of the noise layer.
+    """Opacity of the noise layer.
 
     :attr:`noise_opacity` is a :class:`~kivy.properties.NumericProperty` and
-    defaults to 0.1.'''
+    defaults to 0.1."""
 
-    luminosity = NumericProperty(1.25)
-    '''Luminosity boost that will be aplied to the background.
+    saturation = NumericProperty(1.2)
+    """Saturation boost that will be aplied to the background.
+
+    :attr:`saturation` is a :class:`~kivy.properties.NumericProperty` and
+    defaults to 1.2."""
+
+    luminosity = NumericProperty(1.3)
+    """Luminosity boost that will be aplied to the background.
 
     :attr:`luminosity` is a :class:`~kivy.properties.NumericProperty` and
-    defaults to 1.25.'''
+    defaults to 1.3."""
 
-    overlay_color = ColorProperty([1, 1, 1, 0.75])
-    '''Color/tint overlay that will be aplied to the background.
+    overlay_color = ColorProperty([0.5, 0.5, 0.5, 0.5])
+    """Color/tint overlay that will be aplied over the background.
 
     :attr:`overlay_color` is a :class:`~kivy.properties.ColorProperty` and
-    defaults to [1, 1, 1, 0.6].'''
+    defaults to [0.5, 0.5, 0.5, 0.35]."""
 
     border_radius = ListProperty([0, 0, 0, 0])
-    '''Border radius that will be used by the default canvas shape
-    (RoundedRectangle).
+    """Specifies the radius used for the rounded corners clockwise:
+    top-left, top-right, bottom-right, bottom-left.
 
     :attr:`border_radius` is a :class:`~kivy.properties.ListProperty` and
-    defaults to [0, 0, 0, 0].'''
+    defaults to [0, 0, 0, 0]."""
 
-    downscale_factor = NumericProperty(8)
-    '''Determine how many times the fbo size will be reduced.
+    outline_color = ColorProperty([1.0, 1.0, 1.0, 1.0])
+    """FrostedGlass outline color.
 
-    This property affects the performane directly. As bigger the FBO is,
-    more slow will be the process to compute the gaussian blur resulting in
-    significative fps drop.
+    :attr:`outline_color` is a :class:`~kivy.properties.ColorProperty` and
+    defaults to [1.0, 1.0, 1.0, 1.0]."""
 
-    :attr:`downscale_factor` is a :class:`~kivy.properties.NumericProperty` and
-    defaults to 8.'''
+    outline_width = NumericProperty(1)
+    """FrostedGlass outline width.
+
+    :attr:`outline_width` is a :class:`~kivy.properties.NumericProperty` and
+    defaults to 1."""
 
     def __init__(self, **kwargs):
-        super(FrostedGlass, self).__init__(**kwargs)
-
-        self.canvas = RenderContext(use_parent_projection=True,
-                                    use_parent_modelview=True,
-                                    use_parent_frag_modelview=True)
-        with self.canvas:
-            self.fbo_color = Color(1, 1, 1, 1)
-            self.bt = BindTexture(texture=self.background, index=2)
-            self.fbo_rect = RoundedRectangle(size=self.size,
-                                             pos=self.pos,
-                                             radius=self.border_radius)
-
-        self.canvas.shader.fs = shader_effect
-
-        self.auto_update = Clock.schedule_interval(self.update_shader, 1/30)
-        self.auto_update.cancel()
-
-        self.stop_auto_update_effect = Clock.schedule_once(
-            lambda dt: self.auto_update.cancel(), 2
+        super().__init__(**kwargs)
+        self.bind(
+            blur_size=self.update_effect,
+            noise_opacity=self.update_effect,
+            luminosity=self.update_effect,
+            saturation=self.update_effect,
+            overlay_color=self.update_effect,
+            border_radius=self.update_effect
         )
-        self.stop_auto_update_effect.cancel()
 
-        self.bind(on_kv_post=self.update_effect)
+        self.frosted_glass_effect = FloatLayout()
+        self.frosted_glass_effect.canvas = RenderContext(
+            use_parent_projection=True,
+            use_parent_modelview=True,
+            use_parent_frag_modelview=True
+        )
+        with self.frosted_glass_effect.canvas:
+            self.bt_1 = BindTexture(index=1)
+            self.bt_2 = BindTexture(index=2)
+            self.fbo_rect = RoundedRectangle(
+                size=self.size,
+                pos=self.pos,
+                radius=self.border_radius,
+            )
+        self.frosted_glass_effect.canvas.shader.fs = final_shader_effect
+        self.frosted_glass_effect.canvas.shader.vs = vertex_shader
 
-    def update_effect(self, *args):
-        Clock.schedule_once(self.update_shader, 1/60)
+        self.add_widget(self.frosted_glass_effect)
 
-    def start_auto_update_effect(self):
-        if self.stop_auto_update_effect.is_triggered:
-            self.stop_auto_update_effect.cancel()
-        if not self.auto_update.is_triggered:
-            self.auto_update()
+        with self.canvas:
+            self._outline_color = Color(rgba=self.outline_color)
+            self.outline = SmoothLine(
+                width=1,
+                overdraw_width=dp(1.5),
+                rounded_rectangle=(
+                    self.x, self.y, self.width, self.height, 1, 1, 1, 1
+                )
+            )
 
-    def update_shader(self, *largs):
-        self.fbo_rect.radius = self.border_radius
+        self.horizontal_blur = HorizontalBlur(blur_size=self.blur_size)
+        self.vertical_blur = VerticalBlur(blur_size=self.blur_size)
+        self.noise = Noise()
 
-        self.canvas['opacity'] = float(self.opacity)
-        self.canvas['blur_size'] = float(dp(self.blur_size))
-        self.canvas['luminosity'] = float(self.luminosity)
-        self.canvas['noiseOpacity'] = float(self.noise_opacity)
-        self.canvas['color_overlay'] = [float(x) for x in self.overlay_color]
+        self.last_value = 0
+        self.last_value_list = [0, 0]
+        self.last_update_time = 0
+
+        self.update_fbo_effect()
+
+    def update_effect(self, *args, type=None):
+        Clock.schedule_once(partial(self.update_glsl, type), 0)
+
+    def refresh_effect(self, *args):
+        self.update_fbo_effect()
+        self.update_effect(type=type)
+
+    def update_glsl(self, *args):
+        fge_canvas = self.frosted_glass_effect.canvas
+
+        fge_canvas["luminosity"] = float(self.luminosity)
+        fge_canvas["saturation"] = float(self.saturation)
+        fge_canvas["noise_opacity"] = float(self.noise_opacity)
+        fge_canvas["color_overlay"] = [float(v) for v in self.overlay_color]
 
         pos = self.to_window(*self.pos)
-        self.canvas['resolution'] = [float(v) for v in self.size]
-        self.canvas['position'] = [float(v) for v in pos]
+        fge_canvas["resolution"] = [float(v) for v in self.size]
+        fge_canvas["position"] = [float(v) for v in pos]
 
-        self.fbo = Fbo(size=(self.width/self.downscale_factor,
-                             self.height/self.downscale_factor))
+        if "in_motion" in args:
+            self.update_fbo_effect(True)
+        elif (
+            self.fbo_1.size == (min(self.width, 150), min(self.height, 150))
+            or self.fbo_2.size == (min(self.width, 150), min(self.height, 150))
+        ):
+            self.update_fbo_effect()
 
-        with self.fbo.before:
-            PushMatrix()
-        with self.fbo:
-            ClearColor(0.7, 0.7, 0.7, 1)
-            ClearBuffers()
-            Scale(1/self.downscale_factor)
-            Translate(-pos[0], -pos[1])
-        with self.fbo.after:
-            PopMatrix()
+        self.vertical_blur.fbo = self.fbo_1
+        self.vertical_blur.fbo.add(self.background.canvas)
+        self.vertical_blur.fbo.draw()
+        self.vertical_blur.rect.size = self.size
+        self.vertical_blur.rect.pos = pos
+        self.vertical_blur.rect.texture = self.vertical_blur.fbo.texture
+        self.vertical_blur.fbo.remove(self.background.canvas)
 
-        self.fbo.add(self.background.canvas)
-        self.fbo.draw()
-        self.bt.texture = self.fbo.texture
-        self.canvas['bg'] = 2
-        self.fbo.remove(self.background.canvas)
+        self.horizontal_blur.fbo = self.fbo_2
+        self.horizontal_blur.fbo.add(self.vertical_blur.canvas)
+        self.horizontal_blur.fbo.draw()
+        self.horizontal_blur.fbo.remove(self.vertical_blur.canvas)
 
-    def on_size(self, instance, value):
-        self.fbo_rect.size = value
-        self.update_effect()
-
-    def on_pos(self, instance, value):
-        self.fbo_rect.pos = value
-        self.update_effect()
+        final_texture = self.fbo_2.texture
+        self.bt_1.texture = final_texture
+        fge_canvas["texture1"] = 1
+        fge_canvas.ask_update()
 
     def on_touch_down(self, touch):
         if self.collide_point(*touch.pos):
+            super().on_touch_down(touch)
             return True
+        return super().on_touch_down(touch)
+
+    def on_size(self, instance, size):
+        self.update_canvas()
+        self.update_fbo_effect()
+        self.update_effect()
+        self.update_noise_texture()
+
+    def update_noise_texture(self):
+        self.fbo_noise = Fbo(size=(self.width/dp(1), self.height/dp(1)))
+        self.noise.fbo = self.fbo_noise
+        self.noise.rect.size = self.size
+        self.noise.fbo.add(self.noise.canvas)
+        self.noise.fbo.draw()
+        self.noise.fbo.remove(self.noise.canvas)
+        self.bt_2.texture = self.fbo_noise.texture
+        self.frosted_glass_effect.canvas["texture2"] = 2
+
+    def on_pos(self, *args):
+        self.update_canvas()
+        self.update_fbo_effect()
+        self.update_effect()
+
+    def on_border_radius(self, *args):
+        self.update_canvas()
+
+    def update_canvas(self):
+        border_radius = list(
+            map(
+                lambda x: max(1, min(min(self.width/2, self.height/2), x)),
+                self.border_radius
+            )
+        )
+        self.fbo_rect.size = self.size
+        self.fbo_rect.pos = self.pos
+        self.fbo_rect.radius = border_radius
+
+        self._outline_color.rgba = self.outline_color
+        self.outline.width = self.outline_width
+        self.outline.rounded_rectangle = (
+            self.x, self.y, self.width, self.height,
+            border_radius[3],
+            border_radius[2],
+            border_radius[1],
+            border_radius[0],
+        )
+
+    def on_blur_size(self, instance, blur_size):
+        self.vertical_blur.blur_size = dp(blur_size)
+        self.horizontal_blur.blur_size = dp(blur_size)
+        self.update_fbo_effect()
+        self.update_effect()
+
+    def update_fbo_effect(self, improve_performance=False):
+        if improve_performance:
+            fbo_size = (min(self.width, 150), min(self.height, 150))
+        else:
+            fbo_size = (min(self.width, 250), min(self.height, 250))
+
+        self.fbo_1 = Fbo(size=fbo_size)
+        self.fbo_2 = Fbo(size=fbo_size)
+
+        pos = self.to_window(*self.pos)
+        x = 1/(self.width/fbo_size[0])
+        y = 1/(self.height/fbo_size[1])
+        z = 1
+
+        with self.fbo_1:
+            ClearColor(0, 0, 0, 0)
+            ClearBuffers()
+            Scale(x, y, z)
+            Translate(-pos[0], -pos[1])
+
+        with self.fbo_2:
+            ClearColor(0, 0, 0, 0)
+            ClearBuffers()
+            Scale(x, y, z)
+            Translate(-pos[0], -pos[1])
+
+    def on_background(self, *args):
+        self.bind_parent_properties(self.background)
+        self.bind_children_properties(self.background)
+
+    def on_parent(self, *args):
+        self.bind_parent_properties(self.parent, check_in_motion=True)
+
+    def bind_children_properties(self, widget):
+        children_list = widget.children
+        children_list_temp = []
+        while children_list:
+            for w in children_list:
+                if isinstance(w, ScrollView):
+                    w.bind(
+                        size=self.trigger_update_effect,
+                        pos=self.trigger_update_effect,
+                        scroll_x=self.trigger_update_effect,
+                        scroll_y=self.trigger_update_effect
+                    )
+                elif isinstance(w, Screen):
+                    w.bind(
+                        on_enter=lambda dt: Clock.schedule_once(
+                            self.refresh_effect, 0
+                        )
+                    )
+                elif isinstance(w, Image):
+                    w.bind(
+                        source=lambda instance, _: instance._coreimage.bind(
+                            on_texture=lambda instance:
+                            self.trigger_update_effect(instance, None)
+                        )
+                    )
+
+                else:
+                    w.bind(
+                        size=self.trigger_update_effect,
+                        pos=self.trigger_update_effect
+                    )
+
+                if w.children:
+                    for wi in w.children:
+                        children_list_temp.append(wi)
+            children_list = children_list_temp
+            children_list_temp = []
+
+    def bind_parent_properties(self, widget, check_in_motion=False):
+        while True:
+            if isinstance(widget, ScrollView):
+                widget.bind(
+                    size=partial(
+                        self.trigger_update_effect,
+                        type="in_motion" if check_in_motion else None
+                    ),
+                    pos=partial(
+                        self.trigger_update_effect,
+                        type="in_motion" if check_in_motion else None
+                    ),
+                    scroll_x=partial(
+                        self.trigger_update_effect,
+                        type="in_motion" if check_in_motion else None
+                    ),
+                    scroll_y=partial(
+                        self.trigger_update_effect,
+                        type="in_motion" if check_in_motion else None
+                    )
+                )
+            elif isinstance(widget, Screen):
+                widget.bind(on_enter=self.refresh_effect)
+            else:
+                widget.bind(
+                    size=self.trigger_update_effect,
+                    pos=self.trigger_update_effect
+                )
+
+            if widget.parent:
+                widget = widget.parent
+            else:
+                break
+
+    def trigger_update_effect(self, widget, value, type=None):
+        delta_time = now() - self.last_update_time
+        if value is None:
+            self.update_effect(type=type)
+
+        elif (
+            (isinstance(value, int) or isinstance(value, float))
+            and delta_time >= 0.016 and round(value, 3) != self.last_value
+        ):
+            self.update_effect(type=type)
+            self.last_value = round(value, 3)
+            self.last_update_time = now()
+
+        elif (
+            isinstance(value, list) and delta_time >= 0.016
+            and (
+                round(value[0], 2) != self.last_value_list[0]
+                or round(value[1], 2) != self.last_value_list[1]
+            )
+        ):
+            self.update_effect(type=type)
+            self.last_value_list = round(value[0], 2), round(value[1], 2)
+            self.last_update_time = now()