+++ CHANGELOG.md
@@ -1,4 +1,47 @@
 # Changelog
+<br>
 
-## [v0.1.1rc1] - 2021-10-02
- - Initial pre release.
+# 0.2.0 → 2022-03-08
+
+Highlights
+----------
+
+- Significant improvements in overall performance.
+- Improved blur quality.
+- Automatic effect update to keep it in sync with the background.
+- Improved noise texture quality and consistency of its appearance on different devices.
+
+Added
+----------
+
+- Added `saturation` property.
+- Added `outline_color` property.
+- Added `outline_width` property.
+- Added `refresh_effect()` method.
+- Added feature to add widgets to the container FrostedGlass #1.
+
+Changed
+----------
+
+- Default `luminosity` value changed from `1.25` to `1.3`.
+- Default `overlay_color` value changed from `[1, 1, 1, 0.75]` to `[0.5, 0.5, 0.5, 0.5]`.
+- Default `luminosity` value changed from `1.25` to `1.3`.
+- Default `blur_size` value changed from `60` to `25`.
+- `update_effect()` does not need to be called manually to update the effect. This is already managed by the widget.
+
+Removed
+----------
+
+- Removed `downscale_factor` property.
+- Removed `start_auto_update_effect()` method.
+- Removed `stop_auto_update_effect()` method.
+
+Fixed
+----------
+
+- Temporary FPS drop when there was a touch on the screen (even using static FrostedGlass widgets).
+
+<br>
+
+# 0.1.1rc1 → 2021-10-02
+ - Initial pre-release.